Make ScoreBoardDrawer a subclass of ScoreBoard
